requirements.txt
Flask~=3.1.2
validators~=0.35.0
pandas~=2.3.2
tldextract~=5.3.0
requests~=2.32.5
python-dotenv~=1.1.1